requirements.txt
langchain==0.3.25
langchain_openai==0.3.17
pandas==2.2.3
pydantic==2.11.4
python-dotenv==1.1.0
streamlit==1.45.1
